chore: update namespaces for manifest related code
