+++ lib/metanorma/array_monkeypatch.rb
@@ -1,0 +1,9 @@
+# TODO sidesteps bibdata redefinition of filter, will remove
+
+module Metanorma
+  module ArrayMonkeypatch
+    class ::Array
+      alias filter select
+    end
+  end
+end
