+++ lib/metanorma/collection_config/bibdata.rb
@@ -1,0 +1,11 @@
+require 'relaton'
+require 'shale'
+
+module Metanorma
+  module CollectionConfig
+
+    class Bibdata < ::Shale::Mapper
+      model ::RelatonBib::BibliographicItem
+    end
+  end
+end
