+++ lib/metanorma/collection_files_lookup/files_lookup_sectionsplit.rb
@@ -1,3 +1,5 @@
+require_relative "../sectionsplit"
+
 module Metanorma
   # XML collection renderer
   class FileLookup
@@ -70,7 +72,7 @@ module Metanorma
 
     def sectionsplit(ident)
       file = @files[ident][:ref]
-      @sectionsplit = Sectionsplit
+      @sectionsplit = ::Metanorma::Sectionsplit
         .new(input:  file, base: @files[ident][:out_path], dir: File.dirname(file),
              output: @files[ident][:out_path], compile_opts: @parent.compile_options,
              fileslookup: self, ident: ident, isodoc: @isodoc)
