+++ lib/metanorma/shale_monkeypatch.rb
